# improve invalidStatusLine by  appending a `[]byte` directly