--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@ tags
 *.fasthttp.br
 .idea
 .DS_Store
+vendor/