--- a/status.go
+++ b/status.go
@@ -183,7 +183,13 @@ func invalidStatusLine(statusCode int) []byte {
 	statusText := StatusMessage(statusCode)
 	// xxx placeholder of status code
 	var line = make([]byte, 0, len("HTTP/1.1 xxx \r\n")+len(statusText))
-	line = append(line, []byte("HTTP/1.1 ")...)
+	if statusCode < 0 {
+		line = append(line, []byte("HTTP/1.1 -")...)
+		statusCode = -statusCode
+	} else {
+		line = append(line, []byte("HTTP/1.1 ")...)
+	}
+
 	line = AppendUint(line, statusCode)
 	line = append(line, []byte(" ")...)
 	line = append(line, []byte(statusText)...)