we  create  a  new `[]byte` by   appending ` []byte` directly  instead of  ` fmt.Sprintf `+`[]byte(str)` . it  not only reduce memory allocation ,but also improve performance.
benchmark result:
- old
```
$tyltr > go test -benchmem -run=none  -bench=BenchmarkStatusLine512 -count 10
goos: darwin
goarch: amd64
pkg: github.com/valyala/fasthttp
BenchmarkStatusLine512-8        12142441               109 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8         8737126               126 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8         9828290               108 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8        12083548               106 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8        12042510               101 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8         9378646               116 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8        10912117               119 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8         8671092               144 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8        10802367               112 ns/op             120 B/op          4 allocs/op
BenchmarkStatusLine512-8        10569896               104 ns/op             120 B/op          4 allocs/op
PASS
ok      github.com/valyala/fasthttp     14.813s
```

- New

```
$tyltr > go test -benchmem -run=none  -bench=BenchmarkStatusLine512 -count 10
goos: darwin
goarch: amd64
pkg: github.com/valyala/fasthttp
BenchmarkStatusLine512-8        38009354                31.6 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        33037570                32.2 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        33171657                35.2 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        31569279                34.4 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        33563328                38.0 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        29985741                37.3 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        35844474                35.7 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        33703874                32.2 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        34399201                32.3 ns/op            51 B/op          2 allocs/op
BenchmarkStatusLine512-8        33057207                32.0 ns/op            51 B/op          2 allocs/op
PASS
ok      github.com/valyala/fasthttp     13.049s

```