@erikdubbelboer 2021-09-01
Why stop there, why not append to a ʼ[]byteʼ directly?